Update 2019-09-08-ultralearning.md
Some grammar and punctuation changes. Seeing a lot of "its vs it's"... just remember that "it's" is a contraction of "it is"; then by elimination you'll remember "its" is the possessive. 

There was a section on contrarianism found in "drilling" that wasn't getting anywhere (I think you were trying to tie 'contrariness can waste time' to 'laggers and leaders' by means of 'drill the weakest link', but the actual logical wrapping 'contrarians are laggers' was forgotten someway through), so I moved that into the chapter on "feedback" with some heavy edits, where I thought it was more appropriate. If it's not good there, you should probably nix it entirely and save it for its own blog post (it's an interesting point). 

You have a few shout-outs to 'common knowledge' that aren't common.
>> Most people don't say "wow, that person's memory is so good because of talent" at the memory
expert contests-- we all know it's training.
I was actually pleasantly surprised it was training (and not savant levels of talent). It might seem harmless sometimes to appeal to common knowledge, but when it's not there, it sounds like you're sneaking a claim by your audience, which damages your credibility. Usually it is better to forgo the potential chaff and just lay the claim out there: "memory champions aren't gifted savants-- they train their memory." 

As an aside, I found this one interesting:
https://archive.nytimes.com/www.nytimes.com/interactive/2011/02/20/magazine/mind-secrets.html?hp
I was using some of the techniques in there to card count for Bridge (which I'm still bad at)

Kill the word "basically" and all of its derivatives. It makes an assumption to 'common knowledge' that readers might not have, and makes the sentence more wordy. It will also help trim your writing down (this is a lot of blog to get through in one sitting). When you're tempted to use "basically," consider eliminating the previous sentence with what you're about to write. 

Read what you write aloud once before you post, that will help you catch sentence structure hiccups.